requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
python-dateutil>=2.8.2
pytz>=2021.1

# Data handling and APIs
requests>=2.26.0
websockets>=10.0
ccxt>=2.0.0  # For cryptocurrency exchange support
yfinance>=0.1.63  # For stock market data

# Analysis and ML
scipy>=1.7.0
scikit-learn>=0.24.2
ta>=0.7.0  # Technical analysis library

# Configuration and logging
pyyaml>=5.4.1
python-dotenv>=0.19.0
loguru>=0.5.3

# Testing
pytest>=6.2.5
pytest-cov>=2.12.0

# Development tools
black>=21.6b0
flake8>=3.9.0
isort>=5.9.0

# Jupyter (optional, for analysis)
jupyter>=1.0.0
notebook>=6.4.0
